requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
python-dotenv==1.1.0
firebase-admin==6.8.0
langchain==0.3.25
langchain-core==0.3.58
langchain-anthropic==0.3.12